# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Applai API — alumni leads from resume
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnalysisInput {
  /** Resume PDF file */
  resume: Blob;
  /** Student's selected major */
  major: string;
  /** Student's university (used to find same-school alumni) */
  university: string;
}

export interface ExtractedKeywords {
  skills: string[];
  internships: string[];
  clubs: string[];
}

export interface CareerEntry {
  role: string;
  organization: string;
  /** @nullable */
  startYear?: number | null;
  /** @nullable */
  endYear?: number | null;
  /** @nullable */
  description?: string | null;
}

export interface EducationEntry {
  institution: string;
  /** @nullable */
  degree?: string | null;
  /** @nullable */
  field?: string | null;
  /** @nullable */
  startYear?: number | null;
  /** @nullable */
  endYear?: number | null;
}

export interface Lead {
  id: string;
  name: string;
  /** @nullable */
  photoUrl?: string | null;
  initials?: string;
  graduationYear: number;
  currentRole: string;
  currentOrganization: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  major?: string | null;
  /** 0-100 overlap between student's extracurriculars and lead's early-career roles */
  similarityScore: number;
  /**
   * One-line description of how impressively this person has grown
   * @nullable
   */
  growthSummary?: string | null;
  matchedSkills: string[];
  matchedClubs: string[];
  matchedInternships: string[];
  careerHistory: CareerEntry[];
  educationHistory: EducationEntry[];
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  hasPublicEmail: boolean;
  hasPublicPhone: boolean;
  sourceUrls?: string[];
}

export interface OutreachDraft {
  emailSubject: string;
  emailBody: string;
  linkedinMessage: string;
}

export interface Analysis {
  id: string;
  major: string;
  university: string;
  keywords: ExtractedKeywords;
  leads: Lead[];
  createdAt: string;
}