# lib/api-zod/src/generated/types/outreachDraft.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Applai API — alumni leads from resume
 * OpenAPI spec version: 0.1.0
 */

export interface OutreachDraft {
  emailSubject: string;
  emailBody: string;
  linkedinMessage: string;
}